services/paper-trading-engine/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.1
numpy==1.26.0
python-dotenv==1.0.0
sqlalchemy==2.0.21
requests==2.31.0
